--- pinot-controller/src/main/java/org/apache/pinot/controller/helix/core/realtime/BlockingSegmentCompletionFSMState.java
@@ -1,30 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.pinot.controller.helix.core.realtime;
-
-public enum BlockingSegmentCompletionFSMState {
-    PARTIAL_CONSUMING,  // Indicates that at least one replica has reported that it has stopped consuming.
-    HOLDING,          // the segment has started finalizing.
-    COMMITTER_DECIDED, // We know who the committer will be, we will let them know next time they call segmentConsumed()
-    COMMITTER_NOTIFIED, // we notified the committer to commit.
-    COMMITTER_UPLOADING,  // committer is uploading.
-    COMMITTING, // we are in the process of committing to zk
-    COMMITTED,    // We already committed a segment.
-    ABORTED,      // state machine is aborted. we will start a fresh one when the next segmentConsumed comes in.
-}
